# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API Swagger description",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/register": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin registration",
                "description": "Register a new admin",
                "operationId": "adminRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "is_admin": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin login",
                "description": "Login for admin users",
                "operationId": "adminLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "generated_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/list-order": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List order history",
                "description": "Retrieve a paginated list of order history for admin users",
                "operationId": "listOrderHistory",
                "responses": {
                    "200": {
                        "description": "List of order history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "devices_uuid": {
                                            "type": "string",
                                            "example": "f55f7549-d185-433c-9a53-ffeb9ebc6c61"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "receipt_token": {
                                            "type": "string",
                                            "example": "etsssstst"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-12-04T04:16:38.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-12-04T04:16:38.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/device/login": {
            "post": {
                "tags": [
                    "Device"
                ],
                "summary": "Device login",
                "description": "Login or create a device",
                "operationId": "deviceLogin",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string",
                                        "example": "123e4567-e89b-12d3-a456-426614174001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "generated_uuid_here"
                                        },
                                        "premium_status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "config_info": {
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/purchase/product": {
            "post": {
                "tags": [
                    "Subsciption"
                ],
                "summary": "Purchase a product",
                "description": "Process a product purchase for authenticated users",
                "operationId": "purchaseProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productId": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "receiptToken": {
                                        "type": "string",
                                        "example": "example_receipt_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product purchased successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}